// src/components/ui/TradingViewChart/charting_library.d.ts
export declare type AvailableSaveloadVersions = "1.0" | "1.1";
// ...
export declare type IBasicDataFeed = IDatafeedChartApi & IExternalDatafeed;
export declare type INumberFormatter = IFormatter<number>;
export declare type InputFieldValidator = (value: any) => InputFieldValidatorResult;
export declare type InputFieldValidatorResult = PositiveBaseInputFieldValidatorResult | NegativeBaseInputFieldValidatorResult;
export declare type LanguageCode = "ar" | "zh" | "cs" | "da_DK" | "nl_NL" | "en" | "et_EE" | "fr" | "de" | "el" | "he_IL" | "hu_HU" | "id_ID" | "it" | "ja" | "ko" | "fa" | "pl" | "pt" | "ro" | "ru" | "sk_SK" | "es" | "sv" | "th" | "tr" | "vi" | "no" | "ms_MY" | "zh_TW";
// ...
export declare type ResolutionString = Nominal<string, "ResolutionString">;
// ...
export declare type Timezone = "Etc/UTC" | CustomTimezones;
// ...
export interface AccessList {
	type: "black" | "white";
	tools: AccessListItem[];
}
// ...
export interface ChartingLibraryWidgetOptions {
	/** @deprecated */
	container_id?: string;
	container: HTMLDivElement | null;
	datafeed: IBasicDataFeed | (IBasicDataFeed & IDatafeedQuotesApi);
	interval: ResolutionString;
	symbol?: string;
	auto_save_delay?: number;
	autosize?: boolean;
	debug?: boolean;
	disabled_features?: string[];
	drawings_access?: AccessList;
	enabled_features?: string[];
	fullscreen?: boolean;
	height?: number;
	library_path?: string;
	locale: LanguageCode;
	numeric_formatting?: NumericFormattingParams;
	saved_data?: object;
	saved_data_meta_info?: SavedStateMetaInfo;
	studies_access?: AccessList;
	study_count_limit?: number;
	symbol_search_request_delay?: number;
	timeframe?: string;
	timezone?: "exchange" | Timezone;
	toolbar_bg?: string;
	width?: number;
	charts_storage_url?: string;
	charts_storage_api_version?: AvailableSaveloadVersions;
	client_id?: string;
	user_id?: string;
	load_last_chart?: boolean;
	studies_overrides?: StudyOverrides;
	customFormatters?: CustomFormatters;
	custom_formatters?: CustomFormatters;
	overrides?: Overrides;
	snapshot_url?: string;
	preset?: "mobile";
	time_frames?: TimeFrameItem[];
	custom_css_url?: string;
	custom_font_family?: string;
	favorites?: Favorites;
	save_load_adapter?: IExternalSaveLoadAdapter;
	loading_screen?: LoadingScreenOptions;
	settings_adapter?: ISettingsAdapter;
	theme?: ThemeName;
	compare_symbols?: CompareSymbol[];
	custom_indicators_getter?: (PineJS: PineJS) => Promise<readonly CustomIndicator[]>;
	additional_symbol_info_fields?: AdditionalSymbolInfoField[];
	header_widget_buttons_mode?: HeaderWidgetButtonsMode;
	context_menu?: ContextMenuOptions;
	time_scale?: TimeScaleOptions;
	custom_translate_function?: CustomTranslateFunction;
	symbol_search_complete?: SymbolSearchCompleteOverrideFunction;
	settings_overrides?: Overrides;
}
// ...
export interface CustomFormatters {
	timeFormatter: CustomFormatter;
	dateFormatter: CustomFormatter;
	tickMarkFormatter?: (date: Date, tickMarkType: TickMarkType) => string;
}
// ...
export interface Favorites {
	intervals: ResolutionString[];
	chartTypes: string[];
}
// ...
export interface IChartWidgetApi {
	onDataLoaded(): ISubscription<() => void>;
	onSymbolChanged(): ISubscription<() => void>;
	onIntervalChanged(): ISubscription<(interval: ResolutionString, timeFrameParameters: {
		timeframe?: TimeFrameValue;
	}) => void>;
	onVisibleRangeChanged(): ISubscription<(range: VisibleTimeRange) => void>;
	onChartTypeChanged(): ISubscription<(chartType: SeriesStyle) => void>;
	dataReady(callback: () => void): boolean;
	crossHairMoved(): ISubscription<(params: CrossHairMovedEventParams) => void>;
	setVisibleRange(range: SetVisibleTimeRange, options?: SetVisibleRangeOptions): Promise<void>;
	setSymbol(symbol: string, callback?: () => void): void;
	setResolution(resolution: ResolutionString, callback?: () => void): void;
	setChartType(type: SeriesStyle, callback?: () => void): void;
	resetData(): void;
	executeActionById(actionId: ChartActionId): void;
	getCheckableActionState(actionId: ChartActionId): boolean;
	refreshMarks(): void;
	clearMarks(): void;
	getAllShapes(): EntityInfo[];
	getAllStudies(): EntityInfo[];
	getPriceToBarRatio(): number | null;
	setPriceToBarRatio(ratio: number, options?: UndoOptions): void;
	isPriceToBarRatioLocked(): boolean;
	setPriceToBarRatioLocked(value: boolean, options?: UndoOptions): void;
	getAllPanesHeight(): number[];
	setAllPanesHeight(heights: readonly number[]): void;
	availableZOrderOperations(sources: readonly EntityId[]): AvailableZOrderOperations;
	sendToBack(entities: readonly EntityId[]): void;
	bringToFront(sources: readonly EntityId[]): void;
	bringForward(sources: readonly EntityId[]): void;
	sendBackward(sources: readonly EntityId[]): void;
	/**
	 * @deprecated Use shape/study API instead.
	 * @see {@link getStudyById}
	 * @see {@link getShapeById}
	 */
	setEntityVisibility(entityId: EntityId, isVisible: boolean): void;
	createStudy<TOverrides extends StudyOverrides>(name: string, forceOverlay?: boolean, lock?: boolean, inputs?: Record<string, StudyInputValue>, overrides?: TOverrides, options?: CreateStudyOptions): Promise<EntityId | null>;
	/**
	 * @deprecated Prefer `createStudy` function that relies on named properties for `inputs`.
	 * @see {@link createStudy}
	 */
	createStudy<TStudyInputValue extends StudyInputValue, TOverrides extends StudyOverrides>(name: string, forceOverlay?: boolean, lock?: boolean, inputs?: TStudyInputValue[], overrides?: TOverrides, options?: CreateStudyOptions): Promise<EntityId | null>;
	getStudyById(entityId: EntityId): IStudyApi;
	getSeries(): ISeriesApi;
	createShape<TOverrides extends object>(point: ShapePoint, options: CreateShapeOptions<TOverrides>): EntityId | null;
	createMultipointShape<TOverrides extends object>(points: ShapePoint[], options: CreateMultipointShapeOptions<TOverrides>): EntityId | null;
	getShapeById(entityId: EntityId): ILineDataSourceApi;
	removeEntity(entityId: EntityId, options?: UndoOptions): void;
	removeAllShapes(): void;
	removeAllStudies(): void;
	selection(): ISelectionApi;
	showPropertiesDialog(studyId: EntityId): void;
	createStudyTemplate(options: CreateStudyTemplateOptions): object;
	applyStudyTemplate(template: object): void;
	createOrderLine(options?: UndoOptions): IOrderLineAdapter;
	createPositionLine(options?: UndoOptions): IPositionLineAdapter;
	createExecutionShape(options?: UndoOptions): IExecutionLineAdapter;
	symbol(): string;
	symbolExt(): SymbolExt;
	resolution(): ResolutionString;
	getVisibleRange(): VisibleTimeRange;
	/**
	 * @deprecated Use Price Scale API instead
	 */
	getVisiblePriceRange(): VisiblePriceRange;
	/**
	 * @deprecated Use rightOffset from TimeScale API instead
	 */
	scrollPosition(): number;
	/**
	 * @deprecated Use defaultRightOffset from TimeScale API instead
	 */
	defaultScrollPosition(): number;
	priceFormatter(): INumberFormatter;
	chartType(): SeriesStyle;
	/**
	 * @deprecated Use Timezone API instead
	 * @see {@link getTimezoneApi}
	 */
	setTimezone(timezone: "exchange" | Timezone): void;
	/**
	 * @deprecated Use Timezone API instead
	 * @see {@link getTimezoneApi}
	 */
	getTimezone(): "exchange" | Timezone;
	getTimezoneApi(): ITimezoneApi;
	getPanes(): IPaneApi[];
	exportData(options?: Partial<ExportDataOptions>): Promise<ExportedData>;
	canZoomOut(): boolean;
	zoomOut(): void;
	setZoomEnabled(enabled: boolean): void;
	setScrollEnabled(enabled: boolean): void;
	shapesGroupController(): IShapesGroupControllerApi;
	barTimeToEndOfPeriod(unixTime: number): number;
	endOfPeriodToBarTime(unixTime: number): number;
	getTimeScale(): ITimeScaleApi;
	isSelectBarRequested(): boolean;
	requestSelectBar(): Promise<number>;
	cancelSelectBar(): void;
}
export interface IChartingLibraryWidget {
	headerReady(): Promise<void>;
	onChartReady(callback: EmptyCallback): void;
	onGrayedObjectClicked(callback: (obj: GrayedObject) => void): void;
	onShortcut(shortCut: string | number | (string | number)[], callback: EmptyCallback): void;
	subscribe<EventName extends keyof SubscribeEventsMap>(event: EventName, callback: SubscribeEventsMap[EventName]): void;
	unsubscribe<EventName extends keyof SubscribeEventsMap>(event: EventName, callback: SubscribeEventsMap[EventName]): void;
	chart(index?: number): IChartWidgetApi;
	getLanguage(): LanguageCode;
	setSymbol(symbol: string, interval: ResolutionString, callback: EmptyCallback): void;
	remove(): void;
	closePopupsAndDialogs(): void;
	selectLineTool(linetool: SupportedLineTools): void;
	selectedLineTool(): SupportedLineTools;
	save(callback: (state: object) => void): void;
	load(state: object, extendedData?: SavedStateMetaInfo): void;
	getSavedCharts(callback: (chartRecords: SaveLoadChartRecord[]) => void): void;
	loadChartFromServer(chartRecord: SaveLoadChartRecord): void;
	saveChartToServer(onComplete?: EmptyCallback, onFail?: EmptyCallback, options?: SaveChartToServerOptions): void;
	removeChartFromServer(chartId: string, onCompleteCallback: EmptyCallback): void;
	onContextMenu(callback: (unixTime: number, price: number) => ContextMenuItem[]): void;
	createButton(options?: CreateButtonOptions): HTMLElement;
	createDropdown(params: DropdownParams): Promise<IDropdownApi>;
	showNoticeDialog(params: DialogParams<() => void>): void;
	showConfirmDialog(params: DialogParams<(confirmed: boolean) => void>): void;
	showLoadChartDialog(): void;
	showSaveAsChartDialog(): void;
	symbolInterval(): SymbolIntervalResult;
	mainSeriesPriceFormatter(): INumberFormatter;
	getIntervals(): string[];
	getStudiesList(): string[];
	getStudyInputs(studyName: string): StudyInputInfo[];
	addCustomCSSFile(url: string): void;
	applyOverrides<TOverrides extends StudyOverrides>(overrides: TOverrides): void;
	applyStudiesOverrides(overrides: object): void;
	watchList(): Promise<IWatchListApi>;
	news(): Promise<INewsApi>;
	widgetbar(): Promise<IWidgetbarApi>;
	activeChart(): IChartWidgetApi;
	chartsCount(): number;
	layout(): LayoutType;
	setLayout(layout: LayoutType): void;
	layoutName(): string;
	changeTheme(themeName: ThemeName, options?: ChangeThemeOptions): Promise<void>;
	getTheme(): ThemeName;
	takeScreenshot(): void;
	takeClientScreenshot(options?: Partial<ClientSnapshotOptions>): Promise<HTMLCanvasElement>;
	lockAllDrawingTools(): IWatchedValue<boolean>;
	hideAllDrawingTools(): IWatchedValue<boolean>;
	magnetEnabled(): IWatchedValue<boolean>;
	magnetMode(): IWatchedValue<number>;
	symbolSync(): IWatchedValue<boolean>;
	intervalSync(): IWatchedValue<boolean>;
	crosshairSync(): IWatchedValue<boolean>;
	timeSync(): IWatchedValue<boolean>;
	startFullscreen(): void;
	exitFullscreen(): void;
	undoRedoState(): UndoRedoState;
	navigationButtonsVisibility(): IWatchedValue<VisibilityType>;
	paneButtonsVisibility(): IWatchedValue<VisibilityType>;
	dateFormat(): IWatchedValue<DateFormat>;
}
// ...
export interface IDatafeedQuotesApi {
	getQuotes(symbols: string[], onDataCallback: QuotesCallback, onErrorCallback: (msg: string) => void): void;
	subscribeQuotes(symbols: string[], fastSymbols: string[], onRealtimeCallback: QuotesCallback, listenerGUID: string): void;
	unsubscribeQuotes(listenerGUID: string): void;
}
// ...
export interface ISettingsAdapter {
	initialSettings?: InitialSettingsMap;
	setValue(key: string, value: string): void;
	removeValue(key: string): void;
}
// ...
export interface LoadingScreenOptions {
	foregroundColor?: string;
	backgroundColor?: string;
}
// ...
export interface NumericFormattingParams {
	decimal_sign: string;
}
// ...
export interface Overrides {
	[key: string]: string | number | boolean;
}
// ...
export interface SavedStateMetaInfo {
	uid: number;
	name: string;
	description: string;
}
// ...
export interface StudyOverrides {
	[key: string]: StudyOverrideValueType;
}
// ...
export interface TimeFrameItem {
	text: string;
	resolution: ResolutionString;
	description?: string;
	title?: string;
}
// ...
export interface TradingTerminalWidgetOptions extends ChartingLibraryWidgetOptions {
	brokerConfig?: SingleBrokerMetaInfo;
	broker_config?: SingleBrokerMetaInfo;
	restConfig?: RestBrokerConnectionInfo;
	widgetbar?: WidgetBarParams;
	rss_news_feed?: RssNewsFeedParams;
	news_provider?: GetNewsFunction;
	trading_customization?: TradingCustomization;
	brokerFactory?(host: IBrokerConnectionAdapterHost): IBrokerWithoutRealtime | IBrokerTerminal;
	broker_factory?(host: IBrokerConnectionAdapterHost): IBrokerWithoutRealtime | IBrokerTerminal;
}